algorithm registration

+++ Ship.h
@@ -95,7 +95,7 @@ public:
     void moveContainer(coordinate origin, coordinate dest);
     void updateFreeSpace(int num);
     bool isOnShip(Container &con);
-    std::shared_ptr<Port> getPortByNumber(int num);
+    std::shared_ptr<Port> getPortByNumber(size_t num);
 
 
 };
